requirements.txt
async-generator==1.10
attrs==23.1.0
certifi==2023.5.7
charset-normalizer==3.1.0
contourpy==1.1.0
cycler==0.11.0
DateTime==5.1
exceptiongroup==1.1.1
filelock==3.12.2
fonttools==4.40.0
gitdb==4.0.10
GitPython==3.1.31
google-images-download==2.8.0
h11==0.14.0
html2image==2.0.3
idna==3.4
importlib-resources==5.12.0
Jinja2==3.1.2
joblib==1.2.0
kiwisolver==1.4.4
MarkupSafe==2.1.3
matplotlib==3.7.1
mpmath==1.3.0
networkx==3.1
numpy==1.25.0
opencv-python==4.7.0.72
outcome==1.2.0
packaging==23.1
pandas==2.0.2
Pillow==9.5.0
psutil==5.9.5
pyparsing==3.1.0
PySocks==1.7.1
python-dateutil==2.8.2
pytz==2023.3
PyYAML==6.0
requests==2.31.0
scikit-learn==1.2.2
scipy==1.10.1
seaborn==0.12.2
selenium==4.10.0
six==1.16.0
sklearn==0.0.post5
smmap==5.0.0
sniffio==1.3.0
sortedcontainers==2.4.0
sympy==1.12
thop==0.1.1.post2209072238
threadpoolctl==3.1.0
torch==2.0.1
torchvision==0.15.2
tqdm==4.65.0
trio==0.22.0
trio-websocket==0.10.3
typing_extensions==4.6.3
tzdata==2023.3
ultralytics==8.0.121
urllib3==2.0.3
wsproto==1.2.0
zipp==3.15.0
zope.interface==6.0
